requirements.txt
psycopg2
pyTelegramBotAPI
PyYAML
loguru